refactoring run.py
